Organize Classes
* Create new packages to better organize classes
