+++ RPG Project/src/game/ui/screens/AreaScreen.java
@@ -1,4 +1,4 @@
-package game.ui;
+package game.ui.screens;
 
 import java.awt.Graphics2D;
 import java.awt.image.BufferedImage;
@@ -11,8 +11,8 @@ import java.util.Map;
 import javax.imageio.ImageIO;
 
 import game.Main;
-import game.Tile;
-import game.entities.Enemy;
+import game.entity.Enemy;
+import game.ui.Tile;
 
 /**
  * This class implements IScreen and is used to display an area
