+++ RPG Project/src/game/ui/screens/CharacterCreationScreen.java
@@ -1,4 +1,4 @@
-package game.ui;
+package game.ui.screens;
 
 import java.awt.Color;
 import java.awt.Dimension;
@@ -17,7 +17,7 @@ import javax.swing.border.Border;
 import javax.swing.border.LineBorder;
 
 import game.Main;
-import game.entities.Player;
+import game.entity.Player;
 
 import static javax.swing.SpringLayout.*;
